Bug 1801028 - part 18: Make `CSSEditUtils::GenerateCSSDeclarationsFromHTMLStyle` take `EditorElementStyle` r=m_kato
And also making it and `BuildCSSDeclarations` return pairs of CSS property and
value with an array makes their callers simpler.

Depends on D162517

Differential Revision: https://phabricator.services.mozilla.com/D162518

+++ editor/libeditor/HTMLStyleEditor.cpp
@@ -439,7 +439,7 @@ Result<bool, nsresult> HTMLEditor::ElementIsGoodContainerForTheStyle(
   if (aStyleAndValue.IsCSSEditable(*styledNewSpanElement)) {
     // MOZ_KnownLive(*styledNewSpanElement): It's newSpanElement whose type is
     // RefPtr.
-    Result<int32_t, nsresult> result = CSSEditUtils::SetCSSEquivalentToStyle(
+    Result<size_t, nsresult> result = CSSEditUtils::SetCSSEquivalentToStyle(
         WithTransaction::No, *this, MOZ_KnownLive(*styledNewSpanElement),
         aStyleAndValue, &aStyleAndValue.mAttributeValue);
     if (MOZ_UNLIKELY(result.isErr())) {
@@ -820,7 +820,7 @@ Result<EditorDOMPoint, nsresult> HTMLEditor::SetInlinePropertyOnNodeImpl(
     if (spanStyledElement && aStyleToSet.IsCSSEditable(*spanStyledElement)) {
       // MOZ_KnownLive(*spanStyledElement): It's spanElement whose type is
       // RefPtr.
-      Result<int32_t, nsresult> result = CSSEditUtils::SetCSSEquivalentToStyle(
+      Result<size_t, nsresult> result = CSSEditUtils::SetCSSEquivalentToStyle(
           WithTransaction::Yes, *this, MOZ_KnownLive(*spanStyledElement),
           aStyleToSet, &aStyleToSet.mAttributeValue);
       if (MOZ_UNLIKELY(result.isErr())) {
